chore: use the same function for camalizing in yaml & js
