+++ packages/cubejs-schema-compiler/src/compiler/CubeSymbols.js
@@ -4,6 +4,7 @@ import { camelize } from 'inflection';
 
 import { UserError } from './UserError';
 import { DynamicReference } from './DynamicReference';
+import { camelizeCube } from './utils';
 
 const FunctionRegex = /function\s+\w+\(([A-Za-z0-9_,]*)|\(([\s\S]*?)\)\s*=>|\(?(\w+)\)?\s*=>/;
 const CONTEXT_SYMBOLS = {
@@ -14,38 +15,6 @@ const CONTEXT_SYMBOLS = {
 };
 
 const CURRENT_CUBE_CONSTANTS = ['CUBE', 'TABLE'];
-
-const CUBE_ALIAS_MAPPING = {
-  sql_alias: 'sqlAlias',
-  sql_table: 'sqlTable',
-  refresh_key: 'refreshKey',
-  pre_aggregations: 'preAggregations',
-  rewrite_queries: 'rewriteQueries',
-  data_source: 'dataSource',
-};
-
-const CUBE_MEASURES_ALIAS_MAPPING = {
-  drill_members: 'drillMembers',
-  rolling_window: 'rollingWindow',
-};
-
-const CUBE_DIMENSIONS_ALIAS_MAPPING = {
-  primary_key: 'primaryKey',
-  propagate_filters_to_sub_query: 'propagateFiltersToSubQuery',
-  sub_query: 'subQuery',
-};
-
-const CUBE_PRE_AGGREGATION_ALIAS_MAPPING = {
-  time_dimension: 'timeDimension',
-  partition_granularity: 'partitionGranularity',
-  refresh_key: 'refreshKey',
-  allow_non_strict_date_range_match: 'allowNonStrictDateRangeMatch',
-  use_original_sql_preaggregations: 'useOriginalSqlPreAggregations',
-  scheduled_refresh: 'scheduledRefresh',
-  build_range_start: 'buildRangeStart',
-  build_range_end: 'buildRangeEnd',
-  union_with_source_data: 'unionWithSourceData',
-};
 
 export class CubeSymbols {
   constructor() {
@@ -73,6 +42,7 @@ export class CubeSymbols {
       const cubeDefinition = this.cubeDefinitions[cubeName];
       this.builtCubes[cubeName] = this.createCube(cubeDefinition);
     }
+
     return this.builtCubes[cubeName];
   }
 
@@ -80,6 +50,7 @@ export class CubeSymbols {
     let measures;
     let dimensions;
     let segments;
+
     const cubeObject = Object.assign({
       allDefinitions(type) {
         if (cubeDefinition.extends) {
@@ -150,10 +121,7 @@ export class CubeSymbols {
       errorReporter.error(`${duplicateNames.join(', ')} defined more than once`);
     }
 
-    this.transformAliases(cube, CUBE_ALIAS_MAPPING, false);
-    this.transformAliases(cube.measures, CUBE_MEASURES_ALIAS_MAPPING);
-    this.transformAliases(cube.dimensions, CUBE_DIMENSIONS_ALIAS_MAPPING);
-    this.transformAliases(cube.preAggregations, CUBE_PRE_AGGREGATION_ALIAS_MAPPING);
+    camelizeCube(cube);
 
     this.camelCaseTypes(cube.joins);
     this.camelCaseTypes(cube.measures);
@@ -172,28 +140,6 @@ export class CubeSymbols {
       cube.segments || {},
       cube.preAggregations || {}
     );
-  }
-
-  /**
-   * @private
-   */
-  transformAliases(obj, mapping, nested = true) {
-    if (!obj) {
-      return;
-    }
-
-    if (nested) {
-      for (const field of Object.keys(obj)) {
-        this.transformAliases(obj[field], mapping, false);
-      }
-    } else {
-      for (const field of Object.keys(mapping)) {
-        if (obj[field] !== undefined) {
-          obj[mapping[field]] = obj[field];
-          delete obj[field];
-        }
-      }
-    }
   }
 
   /**
